motionsense: lis2dw12: Add list_activities driver method
This CL adds a list_activities method to the lisdw12 driver which is
required for body detection. Without this method, a null function
pointer is being called which leads to an ISH crash. The HC that
exposes this only gets called following the kernel booting up. So
currently, the crash only happens once per kernel boot which is why
the ISH is not boot looping.

BUG=b:452233585
TEST=Verified that the ISH does not crash following an EC reboot.

Change-Id: Ia38f198f402d6c1b4b41ae6d5c433bca50419600