driver: bmi3xx: Trigger irq_handler when reenabling interrupt
When reenabling interrupt mode for bmi3xx IMU, it is possible to have an
event pending while in force mode that needs to be clear while in
interrupt mode. Since the event was posted while the interrupts were
disabled, reenabing interrupts is not enough to trigger the interrupt
handler. Run the interrupt handler to clear the interrupt status
register and process pending events if any when re-enabling the
interrupts.

BUG=b:455798788
TEST=anakin: data from BMI3xx flow after resume.

Change-Id: I2738081b216d0c17a478f7948b9c558a9d61d0ad