kaladin: support new keyboard matrix
The new matrix differs from the old matrix only in the position of the right alt key.
For compatibility, for both the new and old matrix can be supported simultaneously by setting the right alt scan code to (row 0, col 13).

BUG=b:421781342
TEST=zmake build kaladin
LOW_COVERAGE_REASON=test code is not yet ready b:421825142

Change-Id: Icfddce99ff2cca6c434c91e926a8d9e0fd5b5ec7