zephyr: shim: Don't assert for unmatched thread IDs
There many native Zephyr threads on the EC that are not mapped to a
corresponding legacy task ID in task_get_current() /
thread_id_to_task_id(). We'd like to deprecate the remaining uses of
task ID to make this function unnecessary. But for now, disable this
assertion and return TASK_ID_INVALID for un-mapped threads.

BUG=b:450804537
BUG=b:456237554
TEST=zmake build pujjocento
TEST=./twister -T zephyr/test/tasks

Change-Id: I2071aec8e5dcd3df526e72ec21d99da4615220c6