fpsensor: Rename validate_fp_buffer_offset to validate_buffer_offset
Rename the utility function 'validate_fp_buffer_offset' to the more
generic 'validate_buffer_offset'. This refactoring removes the 'fp_'
prefix to avoid confusion with the specific 'fp_buffer' variable and
correctly reflect that the function is a generic validation routine for
any buffer or buffer offset, not one limited to a specific buffer type
or usage.

BUG=b:450383638
TEST=CQ

Change-Id: I01a0815231ddb55972d27aed768bf42f9c57a7db