drivers/board_id: Add Board ID handler
This CL takes care of storing the board id of RVP read from ioexpander
in CBI under the tag of MODEL_ID, if it is not stored in CBI already
using a handler function. This CL also ensures to write a common
board_get_version() function to retrieve the board version either from
cache or CBI accordingly for boards that uses this driver. Also,
inorder to avoid multiple definitions of this function, they must be
removed from corresponding ocelot boards that uses this common driver.

BUG=b:455072992, b:426139268
LOW_COVERAGE_REASON=To be handled in b:416418970
TEST=Build and Tested on ocelot RVPs with rest of patch stack;

Verified board version and model_id value in CBI.

Change-Id: I73a30bd32c0e8b9d229b09d2be76628909485e33
Signed-off-by: Poornima Tom <poornima.tom@intel.com>