Add capture type to fingerprint sensor context
This change introduces a new field, `current_capture_type`, to the
`fpsensor_context` struct. This field will store the type of image
data obtained from the most recent successful fingerprint sensor
capture.

To support this, a new value `FP_CAPTURE_TYPE_INVALID` is added to
the `fp_capture_type` enum to represent an uninitialized or invalid
state. The `current_capture_type` member is initialized to this
invalid state.

BUG=b:449014224
TEST=CQ

Change-Id: I5f30f06b5318b2afb9420dcce6365152b36630ca