test: Only test locked scenario for fpdownload
The full `test_command_fpdownload` test, which includes a successful
image download in the unlocked state, currently fails when running in
the Renode simulator. This failure stems from the lack of image
acquisition simulation in Renode.

Specifically, the test crashes in `get_image_frame_params` because the
`global_context.current_capture_type` is not updated from its default
`FP_CAPTURE_TYPE_INVALID` state, as the image capture step is
effectively skipped in the simulation.

To unblock Renode testing, this change modifies the test to cover only
the locked scenario. This scenario returns `EC_ERROR_ACCESS_DENIED` and
does not depend on a preceding successful image acquisition.

The test case for the successful download (unlocked scenario) has been
temporarily disabled. Re-enabling this test is tracked in b/460170742.

The test function is renamed to `test_command_fpdownload_locked` to
reflect its reduced scope.

BUG=b:457818548
TEST=./test/run_device_tests.py -b helipilot -f servo_micro -t \
fpsensor_debug

Change-Id: I801a9bfbe4577a276c9b947353728b8b252bdc0f