battery_info: Call hook on battery state of charge change
In builds without CONFIG_CHARGER, the system was not being notified
when the battery's state of charge (SoC) changed. This hook is
necessary for other modules that need to react to SoC updates.

This patch introduces a function, check_battery_change_soc, which
is called from battery_poll_dynamic_info when CONFIG_CHARGER is
not defined.

This function tracks the previous state_of_charge and
display_charge. If either value changes, it calls the
HOOK_BATTERY_SOC_CHANGE hook to notify the rest of the system.

LOW_COVERAGE_REASON=b:422080009 tracking test code
BUG=b:439824946
TEST=zmake build %bluey
TEST=ensure the hook is called on battery SOC change by using the
battfake command

Change-Id: I748807be4044073bbca6c12b6998a768ff094784