power/qcom_exp: Boot system if AC present and battery is low
Implement a wake mechanism that initiates the boot sequence when
connected to AC only if the battery State of Charge (SoC) is below a
specific threshold (currently 75%).

Register HOOK_BATTERY_SOC_CHANGE to monitor SoC levels. If the device
is on AC and discharges below this threshold, explicitly set 'ac_on'
and wake the chipset task to start booting and complete the charging
process.

This patch also adds a `TODO` to update the
BATTERY_STATE_OF_CHARGE_LOWER_THRESHOLD once the ADSP lite is available.

LOW_COVERAGE_REASON=b:422080009 tracking test code
BUG=b:439824946
TEST=zmake build %bluey
TEST=Boot device, connect AC. Shut down the system.
1. Manually set SoC < 70% (battfake 10) and verify the device boots
   back on shutdown.
2. Manually set SoC > 70% (battfake 80) and verify the device doesn't
   boot on shutdown.
TEST=ensure AP doesn't power-on if AC disconnected.
TEST=ensure the power-on reason on the AP PMK logs is on the expected
lines.
> B -   3812713 - PM: xVdd reset
> B -   3813018 - PM: PON by CBLPWR

Change-Id: I21a027e9ae2c28654d58a3a50bd15e71e30047e2