dsp_comms: Increase max retry attempts to 50
This CL changes the default setting of
PLATFORM_EC_DSP_CLIENT_TX_RETRY_MAX from 20 to 50. When the AP goes
into SOiX mode, the EC is put into deep sleep and when it's in this
state it may not receive tablet_mode change notifications from the
ISH. Pujjolo-ish already had this same value of 50 and this CL removes
the pujjolo-ish specific one as this CL is increasing the default
value to 50 as well.

BUG=b:458054926
TEST=Verified that changes between tablet and clamshell mode are
consistently resulting in the AP waking from suspend mode and am not
getting error messages on the ISH console for failure to send the
tablet mode change notification to the EC.

Change-Id: I2c6ac50f853d4dbcf92ed44b126a2e4c6c9e4114