fingerprint: Correct FPC sensor frame size in DTS files
The `frame-size` property for the FPC sensor in the bloonchipper and
helipilot Device Tree Source (DTS) files was previously set to 25600.
This value only accounted for the raw pixel data (160 * 160).

This change updates the `frame-size` to 26260. This new size correctly
includes an additional 660 bytes for header and metadata overhead,
aligning the DTS configuration with the actual buffer size required
by the FPC1025 sensor for image captures.

BUG=b:450381837
TEST=CQ

Change-Id: I193d300781b476903b535117f8d205e26f18f537