fingerprint: fpc1025: Assert real image size consistency
The FPC1025 driver was incorrectly using a build assertion to check for
`FINGERPRINT_SENSOR_FRAME_SIZE` consistency across different sensor
configurations. For this sensor, the consistency check should instead
be applied to the `FINGERPRINT_SENSOR_REAL_IMAGE_SIZE`.

Rename the macro to `FPC1025_ASSERT_IMAGE_REAL_SIZE_CONSISTENT` and
update its implementation to compare the
`FINGERPRINT_SENSOR_REAL_IMAGE_SIZE` between configurations. This
aligns the assertion with the intended design for the FPC1025 sensor.

This consistency is crucial because other parts of the driver, such as
`fpc_sensor_spi_write_read`, assume the real image size is the same
across all configurations and hardcode the use of index 0 when
checking the size. This assertion ensures that assumption holds true.

BUG=b:450381837
TEST=zmake -D build helipilot --clobber
TEST=zmake -D build bloonchipper --clobber

Change-Id: I3592f2b5df11c013f1581c1046626036cc0dda6d