power/mt8186: allow boot on AC insert from hibernate
The previous power-on sequence for the mt8186 chipset specifically
prevented the AP from booting when waking from hibernate via AC insert.

This requirement has been removed from the latest Chromebook requirement.
This commit updates the power sequence to allow the AP to boot in this
scenario, aligning the device behavior with the new spec.

BUG=b:422891443
TEST=manually

Change-Id: I0d8d1035d260d19edbeaff80b5991e9478d6ae73