app/ec: Allow CROS_EC_RO_SIZE to be set from device tree
CROS_EC_RO_SIZE has traditionally represented the size of the EC RO
image which may be smaller than the EC RO flash region size on some ECs.

This configuration was lost with
https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6478555

Now, we're providing an optional `image-size` node to the device tree so
these configurations can be specified.

BUG=b:455688100
TEST=zmake compare-builds -a successful

Change-Id: Iceb98dd064c3441503a4d04aec95aeedc2495ed7