dsp_comms: Add debounce delay for GMR gpio interrupt
This CL modifies dsp_service_gmr_tablet_switch_isr to have it call a
deferred function which uses CONFIG_GMR_SENSOR_DEBOUNCE_US to make it
consistent with how this interrupt was handled for sensors on the
EC. Without this debounce the ISH is getting notified of the GMR GPIO
level immediately, but then this value is latched in
gmr_tablet_switch_isr_handler and this can lead to synchronization
issues between the EC and ISH on the state of the GMR sensor.

BUG=b:453515967
TEST=Verify that tablet/clamshell mode switch is working and that if
the GMR sensor is activated when the device is in clamshell mode,
tablet mode is exited once the sensor goes inactive.

Change-Id: I723640878ee91b5461cd9c4cea1c43534d5bd5db