tablet_mode: Add call to dsp_service in GMR isr handler
This CL removes dsp_service_gmr_tablet_switch_isr. Instead of a
special handler being used when sensors are on the ISH,
dsp_service_hook_tablet_mode_change is exposed as a public API that
can be called from gmr_tablet_switch_isr_handler. This change ensures
that the GMR sensor GPIO is debounced in one location and the debounce
is done prior to notifying the ISH of the GMR sensor state change.

Without this debounce the ISH is getting notified of the GMR GPIO
level immediately, but then this value is latched in
gmr_tablet_switch_isr_handler and this can lead to synchronization
issues between the EC and ISH on the state of the GMR sensor.

BUG=b:453515967
TEST=Verify that tablet/clamshell mode switch is working and that if
the GMR sensor is activated when the device is in clamshell mode,
tablet mode is exited once the sensor goes inactive.
LOW_COVERAGE_REASON=b/453515967 Need to add a test for
dsp_service_gmr_tablet_switch_isr. Need this fix in
R-143.

Change-Id: I723640878ee91b5461cd9c4cea1c43534d5bd5db