pdc: Wait for negotiated_power_level or RDO when setting a new RDO
Sometimes the PDC needs additional time after the EC sends SET_RDO to
enter into a contract with the port partner in the sink entry flow. If
an RDO is sent, transition to a new substate,
"SNK_ATTACHED_WAIT_FOR_CONTRACT", that polls GET_CONNECTOR_STATUS for
the new RDO to take effect --either by seeing the
`negotiated_power_level` bit in the connector status change bits be set,
or the expected RDO position to be reported. A timeout is enforced,
after which the EC gives up and assumes whatever RDO is already active
on the PDC.

This fixes a race condition where the EC reads back an outdated RDO
before the PDC has had enough time to negotiate with the port partner.

Remove a bug where pdc_power_mgmt was faking the current RDO by updating
its cached copy of connector_status immediately after SET_RDO. It is not
valid to assume that the requested RDO is immediately active. We must
wait until GET_CONNECTOR_STATUS reflects the live RDO.

Update unit tests to simulate delayed negotiation in the PDC emulators.
Increase certain timeouts to compensate for the extra time needed to
wait for the RDO to go live.

BUG=b:452400160
TEST=./twister -T zephyr/test/pdc
TEST=Test on RTK and TI DUTs

Change-Id: I964c04b101ef9772d47018968210721ceb102193