tablet_mode: Add HOOK_CHIPSET_RESET to reset_tablet_mode
This CL adds a hook to reset_tablet_mode to clamshell when the AP
reboots. Currently this function is only called on CHIPSET_STARTUP
which works for cold boot or exiting SOiX mode, but does not get
called when the device is reset via an AP reboot.

BUG=b:442991892
TEST=Verfied that after swicthing to clamshell mode while the AP is
booting, that the device starts up in clamshell mode. Also tested that
rebooting the AP only while in tablet mode results in the device going
into tablet mode a few seconds after the kernel is fully booted.

Change-Id: I915ace4d13558645bc329a8c2ad9bb3fab930c27