tablet_mode: Add HOOK_CHIPSET_RESET to reset_tablet_mode
This CL adds a hook to reset_tablet_mode to clamshell when the AP
reboots. Currently this function is only called on CHIPSET_STARTUP
which works for cold boot or exiting SOiX mode, but does not get
called when the device is reset via an AP reboot.

In addition, this CL adds a new TABLET_TRIGGER which is used to
override gmr_sensor_at_360 when set_tablet_mode is called from
reset_tablet_mode which is only used when lid angle and tablet_mode are
controlled by the ISH. This is necessary to handle the case where the
AP is rebooted while the GMR sensor is active, but then the lid angle
is switched back to clamshell before the kernel has fully booted.

BUG=b:442991892
TEST=Verified that after switching to clamshell mode while the AP is
booting, that the device starts up in clamshell mode. Also tested that
rebooting the AP only while in tablet mode results in the device going
into tablet mode a few seconds after the kernel is fully booted.

Change-Id: I915ace4d13558645bc329a8c2ad9bb3fab930c27