drivers/board_id: Add board ID "ready" callback
Add support for an optional "handler" property to "intel,rvp-board-id"
compatible nodes that specifies the function to call to notify the
receiver that the board ID information is available.

This is needed to support cases where the board ID information is
obtained from components that are not ready when the EC boots.

BUG=b:455072992
TEST=booted on fatcatrvp-npcx with rest of patch stack;
	verified board ID related info is present.

Change-Id: Id6624ab864f55bfdf02813a0577c84475f0fc4b2