zephyr: fingerprint: add flash address handling
Add flash address handling logic to the ELAN80-series fingerprint driver.
This change prepares the driver for ELANI80SA, which will be enabled in
a later patch.

BUG=b:450710832
TEST=Built with zmake -l DEBUG build em32f967 --clobber
Cherry-picked chromium:6781020 and verified enrollment and matching
functionality.
LOW_COVERAGE_REASON=b:406324560, it is not possible to test code requiring a private library

Change-Id: I7d1eb1658bb6ff4df14530ac2f9760249acb7e5d