drivers/board_id: Trigger on S5 exit, not entry
Accessing the pca95xx@22 seems to be unstable during S5 entry from G3
callback. The exact reason is unclear; perhaps the chip has not fully
come out of reset. Accessing the chip in the S5 exit to higher callback
seems to solve the problem.

BUG=b:392207109
TEST=booted on fatcatrvp-npcx
LOW_COVERAGE_REASON=Coverage to be handled in b/416418970

Change-Id: I08bba43fa62d3370e2ed809cdd3dddbf7a74e7fd