led: Link policy to pins via phandle
Currently, the LED driver links policy nodes to pin configuration nodes
using a shared led-color string, which maps to an enum, e.g. LED_RED.

However, this limits the number of colors supported. Pre-defining a
large number of colors in a common global enum is unscalable.

This change refactors the binding to use direct DT references:
 * Changes the led-color property in cros-ec,led-policy to a phandle.
 * Updates all led_pins.dtsi to ensure every color node unique.
 * Updates all led_policy.dtsi to reference these nodes via phandle.
 * Updates the shim driver led.c to resolve the destination node.

The led-id property is retained in the pin nodes to allow the driver
to verify that a policy references a pin node belonging to the correct
LED.

BUG=b:462151527
TEST=zmake build -a
TEST=./twister -T zephyr/test/nissa -T zephyr/test/drivers

Cq-Depend: chrome-internal:8840657
Change-Id: I428c4d15abfdf8aef6446a642429e1fa0550da86