power/qcom_exp: Boot system if AC present and battery is discharged by 1%
Implement a wake mechanism that initiates the boot sequence when
connected to AC only if the battery State of Charge (SoC) discharges by
1%.

Register HOOK_BATTERY_SOC_CHANGE to monitor SoC levels. If the device
is on AC and discharges below 1% threshold, explicitly set 'ac_on'
and wake the chipset task to start booting and complete the charging
process.

Register HOOK_CHIPSET_SHUTDOWN to cache the current battery SoC levels
on every chipset shutdown.

Implement a wake mechanism to power on the chipset** when the device is
connected to AC power and the battery State of Charge (SoC) has
discharged by 1% or more since the last clean chipset shutdown.

This ensures the system resumes from a low-power state to continue
the charging process, preventing excessive passive discharge in the
off state.

- Register HOOK_CHIPSET_SHUTDOWN to cache the current battery SoC.
- Register HOOK_BATTERY_SOC_CHANGE to monitor the current SoC.
If the discharge amount exceeds the 1% threshold while AC is present,
the 'ac_on' flag is set and the chipset task is woken to initiate boot.

LOW_COVERAGE_REASON=b:422080009 tracking test code
BUG=b:439824946
TEST=zmake build %bluey
TEST=Boot device, connect AC. Shut down the system.
TEST=Manually set SoC 1% lower to the current threshold using battfake
   and verify the device boots.
TEST=ensure AP doesn't power-on if AC disconnected.
TEST=ensure the power-on reason on the AP PMK logs is on the expected
lines.
> B -   3812713 - PM: xVdd reset
> B -   3813018 - PM: PON by CBLPWR

Change-Id: If920f7336fa7e762f67956b09c287606b0ef0f85