uldren: drop stale pwms node
Best I can tell this has been added as a placeholder, the property has
not been defined in the binding at that level and is currently being
ignored and not usable, but since http://crrev.com/c/7185699 it also
fails the build.

Just drop it, it can always be added back later in a child node if
needed.

BUG=none
TEST=cq dry run

Change-Id: I6999e46a5a7c714f7ad025e1e754f0ef71895796