hostcmd: Don't print "HC 0x0067 err 9" messages
This log line means the AP polled for new events (EC_CMD_GET_NEXT_EVENT)
but none are available (EC_RES_UNAVAILABLE, 9).

This is not necessarily an error and can be spammed in the logs at a
high rate when Depthcharge is querying for keyboard events in firmware
screens, which overwhelms tests scanning the console output and makes it
difficult to see more important messages.

BUG=None
TEST=Run on WCL-RVP

Change-Id: Ied9c479c05854cd1f5c89dec960a81960cf07796