hostcmd: Don't print "HC 0x0067 err 9" messages
This log line means the AP polled for new events (EC_CMD_GET_NEXT_EVENT)
but none are available (EC_RES_UNAVAILABLE, 9).

This is not necessarily an error and can be spammed in the logs at a
high rate when Depthcharge is querying for keyboard events in firmware
screens, which overwhelms tests scanning the console output and makes it
difficult to see more important messages.

The message will still be displayed if the HD debug level is higher than
NORMAL. This can be set by Kconfig or via the EC console with `hcdebug`.

BUG=None
TEST=Run on WCL-RVP, Msg gone by default, appears if `hcdebug every`

Change-Id: Ied9c479c05854cd1f5c89dec960a81960cf07796