[DO NOT MERGE] power/qcom_exp: Enable charging on shutdown if ac attached
If external power (AC) is present when the chipset shutdown is complete,
wake the AP (Application Processor) to allow charging to continue.

This addresses the scenario where the system is shut down but remains
plugged into the charger, ensuring that the charging process is enabled
by booting the AP into charging mode.

LOW_COVERAGE_REASON=b:422080009 tracking test code
BUG=b:439824946
TEST=zmake build %bluey
TEST=ensure AP boots into charging applet whenever the AC is attached

Change-Id: I76a37a42b8e3588cce88fb829647e4dce31d9de8