[DO NOT MERGE] power/qcom_exp: Enable charging on shutdown if ac attached
If external power (AC) is present when the chipset shutdown is complete,
wake the AP (Application Processor) to allow charging to continue.

This addresses the scenario where the system is shut down but remains
plugged into the charger, ensuring that the charging process is enabled
by booting the AP into charging mode.

Caveat / Known Issue:
This patch cannot be merged currently because reboots are non-functional
due to b/429523381. When the AP is in the charging applet and a Power
button press causes an intended reboot, the existing bug makes the EC
perform a *shutdown* instead. resulting in the AP booting back into the
charging applet instead of completing a normal boot, effectively causing
a boot loop behavior into the applet.

LOW_COVERAGE_REASON=b:422080009 tracking test code
BUG=b:439824946
TEST=zmake build %bluey
TEST=ensure AP boots into charging applet whenever the AC is attached

Change-Id: I76a37a42b8e3588cce88fb829647e4dce31d9de8