fpsensor: support dedicated memory areas of fp data
Add support for using a custom memory area for storing templates and a frame.

Such change is needed if that memory needs additional attributes like
"uncachable" or it has to be placed in other region than zephyr,sram for
any reason e.g. there are multiple sram nodes.

BUG=b:436771254
TEST=Add a sram chosen node for the FP data and make sure the buffers are
placed there

Change-Id: I99342569d7b4991006f4aece8b214dc6f4f29550