dsp_client: delay calling gmr isr handler
On ISH boot flow, ISH first tries to fetch the CBI information.
During this it also receives mode information from EC.
If there is a mode change to tablet mode ISH tries to
communicate this new mode to EC. Suspect this might be
creating an issue as ISH parallelly tries to aquire the CBI
information at the same time.

The changes delay the api gmr_tablet_work_handler() invokation
when sensors are not intialized.

Change-Id: I45bfff1ce5465dd3445608a456ad74d3b17e12d5