panic_log: Include panic_printf
Include panic_printf output in panic_log. This will capture more
context around a panic.

This was not included originally to avoid recursion within panic_log.
panic_log does actually print in any of the paths that could result in
recursion. A way to print may be needed for debugging panic_log, but
this can be worked around.

There was also an assumption that tokenized logging would conflict, but
this assumption was wrong.

BUG=None
TEST=Task info captured in panic_log for watchdog on Delbin

Change-Id: I6718ec9650754d6b0b1f1dafabc375897d64d405