fpmcu: et171: Keep global PMP entries unlocked
Enable the Kconfig option `CONFIG_PMP_NO_LOCK_GLOBAL`. This prevents
the firmware from locking the global Physical Memory Protection (PMP)
entries, including the one for the on-chip ROM, during initial setup.

Previously, the RO firmware would lock the PMP entry protecting the ROM
region. This lock persisted across the system jump to the RW firmware.
Consequently, the RW firmware, upon startup, was unable to modify this
PMP entry to establish the correct memory protection for its own
execution region, undermining the intended security isolation.

Example of the issue:
-   Before: RO locks ROM `0x8d --> 0x80001000 - 0x8003a607 R-X LOCKED`
-   Sysjump to RW fails to reconfigure this entry for the RW range.

With this change:
-   RO protects ROM without locking: `0x0d --> 0x80001000 - 0x8003a547 R-X`
-   RW can successfully configure PMP for its region:
    `0x0d --> 0x80042000 - 0x800b89f7 R-X`

This ensures correct PMP configuration and security isolation for both
the RO and RW stages of the firmware.

BUG=b:406946889
TEST=./test/run_device_tests.py -b sanok -t pmp_entries_ro --zephyr;
     flash on ET171 board; ztest run-all
TEST=./test/run_device_tests.py -b sanok -t pmp_entries_rw --zephyr;
     flash on ET171 board; ztest run-all

Change-Id: I19bcd4ad41eb9fa17fa80b324d32297aa5c28d1f