fpmcu: sanok: Enable HW_STACK_PROTECTION
Enable CONFIG_HW_STACK_PROTECTION for the Sanok program. This feature
utilizes the RISCV PMP to protect the irq and thread main stack
boundaries. It introduces a crash during system jump.

The crash occured because the PMP configuration of the RO image
interfered with the stack initialization of the RW image and vice versa.
We now clear all PMP entries via z_riscv_pmp_clear_all() immediately
before the sysjump to ensure a clean state and prevent the stack crash.

BUG=b:406946889
TEST=dump pmp register entries
TEST=rw 0x00027000 (IRQ stack start address) should crash due to Load
     access fault

Change-Id: Ide7369e404f75c70f2621425e58de9d1610da0ea