fpmcu: et171: Define and protect rollback memory region
Enable memory attribute support (CONFIG_MEM_ATTR) for the Sanok board.

To address the need for protecting rollback metadata (b/406946889),
this change defines a specific memory region for it in the device
tree. This region, named "ROLLBACK_REGION", is located at 0x80040000
and is 8KB in size.

This region is protected using the RISC-V Physical Memory Protection (PMP)
unit by configuring it with the DT_MEM_RISCV_TYPE_EMPTY attribute.
This attribute ensures the region has no access permissions (no read,
no write, and no execute), safeguarding the integrity of the rollback data.

BUG=b:406946889
TEST=dump PMP entries and check rollback entry has the correct
     attributes

Change-Id: Ie271db9cfe26fe28120567faf42c6812ebc1bf36