fpmcu: sanok: Define and protect rollback memory region
Enable memory attribute support (CONFIG_MEM_ATTR) for the Sanok board.

To address the need for protecting rollback metadata (b/406946889),
this change defines a specific memory region for it in the device
tree. This region, named "ROLLBACK_REGION", is located at 0x80004000
and is 8KB in size.

The region is configured with the memory attribute
DT_MEM_RISCV_TYPE_IO_R, marking it as read-only I/O memory. This
utilizes the RISC-V PMP unit to enforce access control, safeguarding
the integrity of the rollback data from writes.

BUG=b:406946889
TEST=dump PMP entries and check rollback entry has the correct
     attributes

Change-Id: Ie271db9cfe26fe28120567faf42c6812ebc1bf36