zephyr: flash: npcx: Preserve flash WP status across sysjump
The flash status register values and protection state were not
being preserved when jumping between RO and RW images,
specifically after a reboot following a sysjump. This caused
software write protection to be lost, leading to the system
becoming unexpectedly unlocked.

This change introduces a mechanism to store the write protection
state (SR1, SR2, and the all_protected flag) using the sysjump tag
mechanism before a jump. During flash initialization, if the
EC_RESET_FLAG_SYSJUMP flag is set, the driver restores the saved
write protection state from the jump tag. This ensures that
software write protection settings persist across sysjumps and
subsequent reboots.

In the older Cros-EC firmware, a feature called UMA_LOCK was used.
UMA_LOCK blocks the UMA controller, preventing any SPI NOR flash
command protocol to the flash. This includes reading Status Registers.
Consequently, to know the flash's software write protection status
(stored in SR1 and SR2) while the system was protected, Cros-EC had
to save these register values into RAM before enabling UMA_LOCK.

The UMA_LOCK bit is an early NPCX feature and is no longer considered
necessary for flash protection. It was removed in the Zephyr firmware
implementation. Because Zephyr does not use UMA_LOCK, it does not
face the same restriction of being unable to read Status Registers
when the flash is protected. However, because it was enabled in the
CROS-EC firmware, we need to respose the SR1 and SR2 to retrieve the
SW-WP status when we boot from a CROS-EC RO. We do not need to preserve
them in the Zephyr firmware because the UMA_LOCK bit is no longer used.

This change mirrors the logic from the original EC repository,
found at https://crrev.com/c/4885833.

BUG=b:445467149
TEST=./test/run_device_tests.py -b helipilot -f servo_micro -t \
     system_is_locked_wp_on.* --zephyr
LOW_COVERAGE_REASON=b:445467149, test was verified on an actual board

Change-Id: I26fbf358db84d7848660a32458c54d2b9c3b8938