zephyr: test: Use fingerprint_get_info() for image frame params
We now call fingerprint_get_info() to fetch the image frame parameters
directly from the sensor device. This works for both emulated version of
the fpsensor_debug test and the one that runs on an physical MCU.

This fixes an issue issue where the fpsensor_debug test failed on boards
(e.g. helipilot) because the local hardcoded definitions and macro for
image_frame_params_arr were using fpsensor_sim DT node label that does
not generalize well to boards.

BUG=b:466229465
TEST=./test/run_device_tests.py -b helipilot -f servo_micro -t
     fpsensor_debug --zephyr
TEST=./twister -v -i -T zephyr/test/fingerprint/task -c

Change-Id: Ia58039890d218d86f3ae5abf1d978090f99c32cd