zephyr: fpmcu: add sanok fpmcu project
Add the sanok FPMCU project that is based on ET171 SOC. The project
uses a EC630 fingerprint sensor.

The sanok project uses a et171 as a board. It is done this way, because
there will be more FPMCU projects using ET171 with almost the same
configuration.

The project assumes using QSPI mode to communicate with a flash chip.

BUG=b:392915103
TEST=Build the sanok firmware, flash it on ET171 development board and
make sure FP related functionality works correctly.

Change-Id: I67bdfd3343bfc836f9e32f8e2b065ea8f5787cc0