keyboard_8042_fn: do not emit codes for non mapped keys
Change the Fn logic to not emit any code if pressing Fn together with a
non mapped key. The current behavior is to emit the normal code in this
situation, but other devices I tried seems to not emit any code in this
condition, and some further research seems to confirm that that would be
the expected behavior.

This unfortunately requires tracking what normal keys have been emitted,
costing few bytes of SRAM.

BUG=none
TEST=unit test and test on a brox device

Change-Id: Ib9816ad7d4a1a12a9c5de3509dfec6e8f626489e