extpower: Update host events on init to prevent async state
Previously, during extpower_init, only the battery memmap flag was
updated with the AC state. The corresponding host event was not
triggered. This created a condition where the host's view of the AC
state could be out of sync on boot.

This commit fixes this by calling extpower_update_host_events during
init. This ensures the proper host event mask is set and the AC event
is sent, preventing a missed update and ensuring the host has the
correct state from startup.

BUG=none
TEST=Host successfully reads the correct AC status.
TEST=Toggle the charger connection and reboot the EC to see the host
event init state.
-> On AC connected and EC reboot -> 0x402488
-> On AC disconnected and EC reboot -> 0x402490

Change-Id: I10417d24c5edc32b29f0a17c341d1383d4f89f03