test: utils: Add SCAN_CONSOLE_LINE to parse multi-line output
The existing `SCAN_CONSOLE_CMD` is useful for simple commands, but
fails on multi-line output (like from `cbi`) because it can only match
from the start of the text.

This CL adds a new `SCAN_CONSOLE_LINE` macro that is more flexible. It
first finds a specific line in the command's output, and then runs
`sscanf` on just that line.

BUG=b:463750635
TEST=Used in the following CL to test the `cbi` console command.

Change-Id: Ifcc27953b517e5e717d5e4a4f9d3aaf73eaf020d