program: add a buch of missing zephyr,deferred-init
Add a bunch of missing zephyr,deferred-init, these are devices that
depends on a device that has a zephyr,deferred-init themselves so
they need to be flagged for deferred init as well, a recent Zephyr
priority check script validates that and fails the build if that's not
he case.

That said, it looks like our code base actually relies on these devices
being initialized out of order, so add a short loop to explicitly
initialize these in their original place, so this should be a no-op
effectively.

BUG=b:320527549
TEST=cq dry run
TEST=run on brox

Change-Id: Iade543eac1b656fa89c1996e3ea700d092025442